# public/swagger/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "User Management API",
        "description": "API for managing users and their profiles",
        "contact": {
            "email": "your.email@example.com"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080/api",
            "description": "Development API Server"
        }
    ],
    "paths": {
        "/profiles": {
            "get": {
                "tags": [
                    "Profiles"
                ],
                "summary": "Get a list of all user profiles",
                "operationId": "066150067037a2eb9501ef5ae97a94af",
                "responses": {
                    "200": {
                        "description": "A list of user profiles",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Profile"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Profiles"
                ],
                "summary": "Create a new user profile",
                "operationId": "e8b73c6c87a7d7cef84ff0805a9c3789",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/NewProfileRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Profile created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Profile"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation errors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/profiles/{id}": {
            "get": {
                "tags": [
                    "Profiles"
                ],
                "summary": "Get a specific user profile by ID",
                "operationId": "a1ef0d100df689e74aa9392ba840038b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the profile to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Details of the requested profile",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Profile"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Profile not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Profiles"
                ],
                "summary": "Update an existing user profile",
                "operationId": "6acf3a201daa330d0f9511bf7d76a2d3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the profile to update",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateProfileRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Profile updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Profile"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation errors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Profile not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Profiles"
                ],
                "summary": "Delete a user profile",
                "operationId": "a7304a5aab130b19dcd19bf3ca4a5ac0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the profile to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Profile deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "error": {
                                            "type": "null",
                                            "example": null
                                        },
                                        "messages": {
                                            "properties": {
                                                "success": {
                                                    "type": "string",
                                                    "example": "Profile deleted successfully."
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Profile not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Profile": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "first_name": {
                        "type": "string",
                        "example": "John",
                        "nullable": true
                    },
                    "last_name": {
                        "type": "string",
                        "example": "Doe",
                        "nullable": true
                    },
                    "website": {
                        "type": "string",
                        "example": "https://example.com",
                        "nullable": true
                    },
                    "instagram": {
                        "type": "string",
                        "example": "johndoe_ig",
                        "nullable": true
                    },
                    "facebook": {
                        "type": "string",
                        "example": "https://facebook.com/johndoe",
                        "nullable": true
                    },
                    "linkedin": {
                        "type": "string",
                        "example": "https://linkedin.com/in/johndoe",
                        "nullable": true
                    },
                    "twitter_x": {
                        "type": "string",
                        "example": "@johndoe_tw",
                        "nullable": true
                    },
                    "avatar": {
                        "type": "string",
                        "example": "profile.jpg",
                        "nullable": true
                    },
                    "bio": {
                        "type": "string",
                        "example": "A short bio about John Doe.",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "ValidationError": {
                "properties": {
                    "errors": {
                        "properties": {
                            "username": {
                                "type": "array",
                                "items": {
                                    "type": "string",
                                    "example": "The username field is required."
                                }
                            },
                            "email": {
                                "type": "array",
                                "items": {
                                    "type": "string",
                                    "example": "The email field must be a valid email address."
                                }
                            },
                            "password": {
                                "type": "array",
                                "items": {
                                    "type": "string",
                                    "example": "The password field is required."
                                }
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}